+++ lib/just_color_picker.dart
@@ -2,6 +2,7 @@
 export 'src/models/color_picker_type.dart';
 export 'src/models/color_state.dart';
 export 'src/models/color_format.dart';
+export 'src/models/color_picker_input_theme.dart';
 
 // Widgets
 export 'src/widgets/color_picker.dart';
@@ -12,6 +13,9 @@ export 'src/widgets/hue_bar.dart';
 export 'src/widgets/hex_input.dart';
 export 'src/widgets/color_preview.dart';
 export 'src/widgets/color_info_panel.dart';
+export 'src/widgets/color_text_field.dart';
+export 'src/widgets/rgb_input.dart';
+export 'src/widgets/hsl_input.dart';
 
 // Utils
 export 'src/utils/color_conversions.dart';